requirements.txt
anthropic
playwright
python-dotenv
pyyaml
